regions: one constructor builds a nested lambda's blueprint
The bytecode emitter and the JIT each hand-wrote the same twenty-field literal
at their `MakeClosure`. `TemplateProto::nested_lambda` is that literal once, so
a field is no longer a backend's to remember — the two tables an error exit
walks least of all.

Three repairs found while reading the sites. The JIT emitted every closure of
the module twice, dropping the first pass whole; the self-tail-call arm
described a slab-pool rotation it no longer emits; and the emitter's
instruction dispatch is split at the destructuring reads to bring it inside
the reading budget, which also gives its two identical struct-key matches one
home.
