+++ src/lir/mod.rs
@@ -1,7 +1,7 @@
-//! Low-level Intermediate Representation (LIR)
-//!
-//! LIR is SSA form with basic blocks and virtual registers.
-//! It is close to the target but still architecture-independent.
+// audited: 2026-09-05
+// docs/impl/lir.md
+//! Low-level Intermediate Representation: SSA form with basic blocks and
+//! virtual registers, close to the target but architecture-independent.
 //!
 //! Pipeline:
 //! ```text
@@ -17,7 +17,7 @@ pub(crate) mod testkit;
 mod types;
 
 pub use display::terminator_kind;
-pub use emit::Emitter;
+pub use emit::{ClosureCompiled, Emitter};
 pub use lower::Lowerer;
 pub use types::{
     closure_value_const_count, for_each_def, for_each_terminator_use, for_each_use,
